Re-organize compose mod
