+++ rustfst/src/algorithms/compose/compose_filters/null_compose_filter.rs
@@ -3,20 +3,20 @@ use std::rc::Rc;
 
 use failure::Fallible;
 
-use crate::algorithms::compose_filters::ComposeFilter;
-use crate::algorithms::filter_states::{FilterState, TrivialFilterState};
-use crate::algorithms::matchers::{MatchType, Matcher};
+use crate::algorithms::compose::compose_filters::ComposeFilter;
+use crate::algorithms::compose::filter_states::{FilterState, TrivialFilterState};
+use crate::algorithms::compose::matchers::{MatchType, Matcher};
 use crate::semirings::Semiring;
-use crate::{Arc, EPS_LABEL};
+use crate::{Arc, NO_LABEL};
 
 #[derive(Debug)]
-pub struct NoMatchComposeFilter<M1, M2> {
+pub struct NullComposeFilter<M1, M2> {
     matcher1: Rc<RefCell<M1>>,
     matcher2: Rc<RefCell<M2>>,
 }
 
 impl<'fst1, 'fst2, W: Semiring + 'fst1 + 'fst2, M1: Matcher<'fst1, W>, M2: Matcher<'fst2, W>>
-    ComposeFilter<'fst1, 'fst2, W> for NoMatchComposeFilter<M1, M2>
+    ComposeFilter<'fst1, 'fst2, W> for NullComposeFilter<M1, M2>
 {
     type M1 = M1;
     type M2 = M2;
@@ -30,14 +30,10 @@ impl<'fst1, 'fst2, W: Semiring + 'fst1 + 'fst2, M1: Matcher<'fst1, W>, M2: Match
     ) -> Fallible<Self> {
         Ok(Self {
             matcher1: m1.into().unwrap_or_else(|| {
-                Rc::new(RefCell::new(
-                    Self::M1::new(fst1, MatchType::MatchOutput).unwrap(),
-                ))
+                Rc::new(RefCell::new(M1::new(fst1, MatchType::MatchOutput).unwrap()))
             }),
             matcher2: m2.into().unwrap_or_else(|| {
-                Rc::new(RefCell::new(
-                    Self::M2::new(fst2, MatchType::MatchInput).unwrap(),
-                ))
+                Rc::new(RefCell::new(M2::new(fst2, MatchType::MatchInput).unwrap()))
             }),
         })
     }
@@ -51,9 +47,12 @@ impl<'fst1, 'fst2, W: Semiring + 'fst1 + 'fst2, M1: Matcher<'fst1, W>, M2: Match
     }
 
     fn filter_arc(&mut self, arc1: &mut Arc<W>, arc2: &mut Arc<W>) -> Fallible<Self::FS> {
-        Ok(Self::FS::new(
-            arc1.olabel != EPS_LABEL || arc2.ilabel != EPS_LABEL,
-        ))
+        let res = if arc1.olabel == NO_LABEL || arc2.ilabel == NO_LABEL {
+            Self::FS::new_no_state()
+        } else {
+            Self::FS::new(true)
+        };
+        Ok(res)
     }
 
     fn filter_final(&self, _w1: &mut W, _w2: &mut W) -> Fallible<()> {
