+++ rustfst/src/algorithms/compose/filter_states/integer_filter_state.rs
