+++ rustfst/src/algorithms/compose/filter_states/pair_filter_state.rs
