+++ rustfst/src/algorithms/compose/filter_states/trivial_filter_state.rs
