+++ rustfst/src/algorithms/compose/lookahead_filters/lookahead_selector.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 
 use failure::_core::cell::RefCell;
 
-use crate::algorithms::matchers::{MatchType, Matcher};
+use crate::algorithms::compose::matchers::{MatchType, Matcher};
 use crate::semirings::Semiring;
 
 #[derive(Debug)]
