+++ rustfst/src/algorithms/compose/lookahead_filters/push_weights_compose_filter.rs
@@ -4,13 +4,13 @@ use std::rc::Rc;
 
 use failure::Fallible;
 
-use crate::algorithms::compose_filters::ComposeFilter;
-use crate::algorithms::filter_states::{FilterState, PairFilterState, WeightFilterState};
-use crate::algorithms::lookahead_filters::lookahead_selector::{MatchTypeTrait, Selector};
-use crate::algorithms::lookahead_filters::LookAheadComposeFilterTrait;
-use crate::algorithms::lookahead_matchers::LookaheadMatcher;
-use crate::algorithms::matchers::MatcherFlags;
-use crate::algorithms::matchers::{MatchType, Matcher};
+use crate::algorithms::compose::compose_filters::ComposeFilter;
+use crate::algorithms::compose::filter_states::{FilterState, PairFilterState, WeightFilterState};
+use crate::algorithms::compose::lookahead_filters::lookahead_selector::{MatchTypeTrait, Selector};
+use crate::algorithms::compose::lookahead_filters::LookAheadComposeFilterTrait;
+use crate::algorithms::compose::lookahead_matchers::LookaheadMatcher;
+use crate::algorithms::compose::matchers::MatcherFlags;
+use crate::algorithms::compose::matchers::{MatchType, Matcher};
 use crate::semirings::{DivideType, Semiring, WeaklyDivisibleSemiring, WeightQuantize};
 use crate::{Arc, KDELTA};
