+++ rustfst/src/algorithms/compose/lookahead_matchers/trivial_lookahead_matcher.rs
