+++ rustfst/src/algorithms/compose/matchers/generic_matcher.rs
@@ -1,4 +1,4 @@
-use crate::algorithms::matchers::SortedMatcher;
+use crate::algorithms::compose::matchers::SortedMatcher;
 
 // TODO: Change this to use InitMatcher once supported.
 pub type GenericMatcher<'fst, F> = SortedMatcher<'fst, F>;
