+++ rustfst/src/algorithms/visitors/mod.rs
@@ -1,4 +1,2 @@
-mod interval_reach_visitor;
 mod scc_visitors;
-pub(crate) use interval_reach_visitor::IntervalReachVisitor;
 pub(crate) use scc_visitors::SccVisitor;
